+++ lib/cequel/type.rb
@@ -90,7 +90,7 @@ module Cequel
         quote(value.utc.to_time)
       when Date
         quote(Time.gm(value.year, value.month, value.day))
-      when Numeric, true, false, Cql::Uuid
+      when Numeric, true, false, Cassandra::Uuid
         value.to_s
       else
         quote_string(value.to_s)
@@ -414,11 +414,11 @@ module Cequel
       end
 
       def cast(value)
-        if value.is_a? Cql::Uuid then value
+        if value.is_a? Cassandra::Uuid then value
         elsif defined?(SimpleUUID::UUID) && value.is_a?(SimpleUUID::UUID)
-          Cql::Uuid.new(value.to_i)
+          Cassandra::Uuid.new(value.to_i)
         elsif value.is_a?(::Integer) || value.is_a?(::String)
-          Cql::Uuid.new(value)
+          Cassandra::Uuid.new(value)
         else
           fail ArgumentError,
                "Don't know how to cast #{value.inspect} to a UUID"
@@ -439,7 +439,7 @@ module Cequel
     #
     class Timeuuid < Uuid
       def cast(value)
-        Cql::TimeUuid.new(super.value)
+        Cassandra::TimeUuid.new(super.value)
       end
 
       def internal_names