+++ spec/examples/metal/keyspace_spec.rb
@@ -98,7 +98,7 @@ describe Cequel::Metal::Keyspace do
     context "with a connection error" do
       it "reconnects to cassandra with a new client after first failed connection" do
         allow(cequel.client).to receive(:execute)
-          .with(statement, cequel.default_consistency)
+          .with(statement, :consistency => cequel.default_consistency)
           .and_raise(Ione::Io::ConnectionError)
           .once